tweak columns acquision order
This allows us again just to reload table data after we restart server,
since we don't require columns (which isn't saved until user selects
save for the first time) but headers which is there.
However, if we do have columns in session, we'll use that (which is
required for code eval to add new columns).